move StopWords decorator to a root level
