+++ src/rooms/dto/create-room.dto.ts
@@ -1,5 +1,5 @@
 import { IsNotEmpty } from 'class-validator';
-import { StopWords } from '../decorators/StopWords';
+import { StopWords } from '../../decorators/StopWords';
 
 export class CreateRoomDto {
   @IsNotEmpty()
